requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1

# Telegram Bot
python-telegram-bot==20.7

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# AI
anthropic==0.42.0

# Recipe scraping
httpx==0.28.1
beautifulsoup4==4.12.3

# Scheduler
apscheduler==3.10.4

# Settings
pydantic-settings==2.7.1
